Add ms8607 PTH sensor support

+++ components/tc74/tc74.h
@@ -4,4 +4,4 @@
 #include "driver/i2c_master.h"
 
 esp_err_t tc74_init(i2c_master_bus_handle_t bus);
-esp_err_t tc74_read_temperature(int8_t *temp_out);
+esp_err_t tc74_read_temperature(float *temp_out);
